# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "مسیر کانکور — API",
    "description": "API بک‌اند سایت آموزشگاه کانکور. اکثر مسیرها نیازمند توکن JWT هستند (هدر Authorization: Bearer <token>).",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "نسبی — همان دامنه‌ی فرانت‌اند (پشت پروکسی/nginx)" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "message": { "type": "string" } }
      },
...
      "Question": {
        "type": "object",
        "required": ["question_text", "options", "correct_index"],
        "properties": {
          "question_text": { "type": "string" },
          "options": { "type": "array", "items": { "type": "string" }, "minItems": 2 },
          "correct_index": { "type": "integer" }
        }
      },
...
    },
    "responses": {
...
      "Forbidden": {
        "description": "نقش کاربر اجازه‌ی این عملیات را نمی‌دهد.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/tests/{id}": {
      "get": {
        "tags": ["tests"],
        "summary": "سوالات یک آزمون (بدون پاسخ صحیح)، نیازمند دسترسی به کورس",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/tests/{id}/submit": {
      "post": {
        "tags": ["tests"],
        "summary": "ارسال پاسخ‌های آزمون و دریافت نمره (فقط شاگرد)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["answers"],
                "properties": {
                  "answers": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": { "question_id": { "type": "integer" }, "selected_index": { "type": "integer" } }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "نتیجه‌ی آزمون" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
    "/tests": {
      "post": {
        "tags": ["tests"],
        "summary": "ساخت آزمون جدید برای یک کورس (استاد/ادمین)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "course_id", "questions"],
                "properties": {
                  "title": { "type": "string" },
                  "course_id": { "type": "integer" },
                  "questions": { "type": "array", "items": { "$ref": "#/components/schemas/Question" }, "minItems": 1 }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "ساخته شد" }, "403": { "$ref": "#/components/responses/Forbidden" } }
      }
    },
...
  }
}